Further work on the API
Added logger to the Product Controller
Added more swagger documentation to the products
Added the dependencies org.springframework.boot, com.mysql,  org.springframework.boot for Database connection.
Added whats needed in applications.proerties to connect to database (not fild inn the correct info in the variables)
Added the necessary notations to the products for the database connection to work.
Created a interface ProductRepository. (not filed with anything.
Cleand up som bad code.